Inline TimeClassExecution within ExecuteLifecycle, now that it is only used once.
