+++ src/Fixie/Execution/Behaviors/ExecuteLifecycle.cs
@@ -1,8 +1,9 @@
-namespace Fixie.Execution.Behaviors
+﻿namespace Fixie.Execution.Behaviors
 {
     using System;
     using System.Collections.Generic;
     using System.Diagnostics;
+    using System.Linq;
 
     class ExecuteLifecycle
     {
@@ -15,26 +16,50 @@ namespace Fixie.Execution.Behaviors
 
         public void Execute(Type testClass, IReadOnlyList<Case> cases)
         {
-            var timeClassExecution = new TimeClassExecution();
+            var stopwatch = Stopwatch.StartNew();
 
-            timeClassExecution.Execute(cases, () =>
+            try
             {
-                try
+                lifecycle.Execute(testClass, caseLifecycle =>
                 {
-                    lifecycle.Execute(testClass, caseLifecycle =>
-                    {
-                        ExecuteCases(caseLifecycle, cases);
-                    });
-                }
-                catch (Exception exception)
-                {
-                    foreach (var @case in cases)
-                        @case.Fail(exception);
-                }
-            });
+                    ExecuteCases(cases, caseLifecycle);
+                });
+            }
+            catch (Exception exception)
+            {
+                foreach (var @case in cases)
+                    @case.Fail(exception);
+            }
+
+            stopwatch.Stop();
+
+            AdjustCaseDurations(cases, stopwatch);
         }
 
-        void ExecuteCases(CaseAction caseLifecycle, IReadOnlyList<Case> cases)
+        static void AdjustCaseDurations(IReadOnlyList<Case> cases, Stopwatch stopwatch)
+        {
+            var classExecutionDuration = stopwatch.Elapsed;
+
+            var totalCaseDuration = TimeSpan.FromTicks(cases.Sum(x => x.Duration.Ticks));
+
+            // Due to the Stopwatch's precision, it is possible that the sum of multiple
+            // imprecise Case measurements will exceed the single imprecise measurement
+            // of the whole class's execution. If so, there's no need to adjust timings.
+
+            if (classExecutionDuration > totalCaseDuration)
+            {
+                var buildChainDuration = classExecutionDuration - totalCaseDuration;
+
+                var numberOfCases = cases.Count;
+
+                var buildChainDurationPerCase = TimeSpan.FromTicks(buildChainDuration.Ticks / numberOfCases);
+
+                foreach (var @case in cases)
+                    @case.Duration += buildChainDurationPerCase;
+            }
+        }
+
+        static void ExecuteCases(IReadOnlyList<Case> cases, CaseAction caseLifecycle)
         {
             foreach (var @case in cases)
             {
